+++ src/init.c
@@ -7,6 +7,13 @@
 
 #include "ps-internal.h"
 #include "common.h"
+#include "config.h"
+
+#ifdef PS__MACOS
+#include <mach/mach_time.h>
+extern struct mach_timebase_info PS_MACH_TIMEBASE_INFO;
+#endif
+
 
 static const R_CallMethodDef callMethods[]  = {
   /* System api */
@@ -21,6 +28,7 @@ static const R_CallMethodDef callMethods[]  = {
   { "ps__tty_size",           (DL_FUNC) ps__tty_size,           0 },
   { "ps__disk_partitions",    (DL_FUNC) ps__disk_partitions,    1 },
   { "ps__disk_usage",         (DL_FUNC) ps__disk_usage,         1 },
+  { "ps__fs_info",            (DL_FUNC) ps__fs_info,            2 },
   { "ps__system_memory",      (DL_FUNC) ps__system_memory,      0 },
   { "ps__system_swap",        (DL_FUNC) ps__system_swap,        0 },
 
@@ -109,6 +117,10 @@ void R_init_ps(DllInfo *dll) {
 
 #ifdef PS__POSIX
   ps__main_thread = pthread_self();
+#endif
+
+#ifdef PS__MACOS
+  mach_timebase_info(&PS_MACH_TIMEBASE_INFO);
 #endif
 
   R_registerRoutines(dll, NULL, callMethods, NULL, NULL);